[PR c++/84729] convert new init to array elt type
A parenthesized initializer is only accepted when new()ing an array in
permissive mode.  We were not careful, however, to convert the
TREE_LIST initializer to the array element type in this extension.
This patch fixes it: after turning the TREE_LIST initializer to a
compound_expr, we convert it to the base type.


for  gcc/cp/ChangeLog

	PR c++/84729
	* init.c (build_vec_init): Convert tree list to base type.

for  gcc/testsuite/ChangeLog

	PR c++/84729
	* g++.dg/pr84729.C: New.